Move pip install code from StartJupyterNotebookAction to WebServer
